% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcPeriodPerformance.R
\name{calcPeriodPerformance}
\alias{calcPeriodPerformance}
\title{Calculating a performance table}
\usage{
calcPeriodPerformance(R)
}
\arguments{
\item{R}{A zoo-object with returns}
}
\value{
Returns a data.frame object showing the instrument and their respective cumulated performance.
}
\description{
This function calculates and displays a table of period performance for financial instruments. A cumulated performance per period is calculated for the different instruments provided. For this purpose, the return data should be log-returns, otherwise the aggregation gives meaningless results.
}
\details{
Currently only an aggregation by year is implemented.
}
\examples{
library(PerformanceAnalytics)
data(edhec)
calcPeriodPerformance(R=edhec[,1:5])
}
\keyword{financial}
\keyword{performance}

