% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financialSummary.R
\name{financialSummary}
\alias{financialSummary}
\title{Create a Summary of Financial Data}
\usage{
financialSummary(deposits, spending)
}
\arguments{
\item{deposits}{A data frame containing deposit transactions with columns:
`Category` and `Amount`.}

\item{spending}{A data frame containing spending transactions with columns:
`Merchant` and `Amount`.}
}
\value{
Prints a summary including total income, total expenses, net balance,
  and aggregated amounts by category and merchant.
}
\description{
This function generates a summary of financial data, including total income,
total expenses, net balance, and aggregated amounts by category and merchant.
}
\keyword{internal}
