% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotPerfMatrix}
\alias{plotPerfMatrix}
\title{Plotting a performance matrix}
\usage{
plotPerfMatrix(r, short = 5, mid = 20, long = 250)
}
\arguments{
\item{r}{A zoo-object with returns}

\item{short}{Lookback period for the shortest time window}

\item{mid}{Lookback period for the middle time window}

\item{long}{Lookback period for the longest time window}
}
\description{
This function plots the performances of financial instruments in the form of a four-field-matrix
}
\examples{
test <- zoo(matrix(rnorm(n=5*300,mean = 0, sd=0.01),300), as.Date("2013-01-01")+0:299)
colnames(test) <- c("Stock A", "Stock B", "Stock C", "Stock D", "Stock E")
plotPerfMatrix(test, short=10, mid=50, long=100)
}
\keyword{financial}
\keyword{performance}

