% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-spending.R
\docType{data}
\name{spending}
\alias{spending}
\title{spending Dataset}
\format{
A data frame with 5743 rows and 5 variables:
\describe{
  \item{Date}{Date in format YYYY-MM-DD}
  \item{Merchant}{Where the transaction was made}
  \item{Category}{Spending category}
  \item{Note}{Note regarding transaction}
  \item{Amount}{Transaction total}
}
}
\source{
Compiled from my personal finance data
}
\usage{
spending
}
\description{
A dataset containing transaction data combined from: credit card statements,
a checking account statement, and payroll deductions.
This data is an output of the statementCombine() function.
}
\keyword{datasets}
